Extract classes to own files, more StorageClient initialization
