+++ google-cloud-storage/src/main/java/com/google/cloud/storage/GapicUnbufferedReadableByteChannel.java
@@ -232,83 +232,3 @@ final class GapicUnbufferedReadableByteChannel
   }
 }
 
-/**
- * Shrink wraps a beginning, offset and limit for tracking state of an individual invocation of
- * {@link #read}
- */
-final class ReadCursor {
-  private final long beginning;
-  private long offset;
-  private final long limit;
-
-  ReadCursor(long beginning, long limit) {
-    this.limit = limit;
-    this.beginning = beginning;
-    this.offset = beginning;
-  }
-
-  public boolean hasRemaining() {
-    return limit - offset > 0;
-  }
-
-  public void advance(long incr) {
-    checkArgument(incr >= 0);
-    offset += incr;
-  }
-
-  public long read() {
-    return offset - beginning;
-  }
-
-  @Override
-  public String toString() {
-    return String.format("ReadCursor{begin=%d, offset=%d, limit=%d}", beginning, offset, limit);
-  }
-}
-
-interface ResponseContentLifecycleManager {
-  ResponseContentLifecycleHandle get(ReadObjectResponse response);
-
-  static ResponseContentLifecycleManager noop() {
-    return response ->
-        new ResponseContentLifecycleHandle(
-            response,
-            () -> {
-              // no-op
-            });
-  }
-}
-
-final class ResponseContentLifecycleHandle implements Closeable {
-  @Nullable private final Closeable dispose;
-
-  private final List<ByteBuffer> buffers;
-
-  ResponseContentLifecycleHandle(ReadObjectResponse resp, @Nullable Closeable dispose) {
-    this.dispose = dispose;
-
-    this.buffers = resp.getChecksummedData().getContent().asReadOnlyByteBufferList();
-  }
-
-  void copy(ReadCursor c, ByteBuffer[] dsts, int offset, int length) {
-    for (ByteBuffer b : buffers) {
-      long copiedBytes = Buffers.copy(b, dsts, offset, length);
-      c.advance(copiedBytes);
-      if (b.hasRemaining()) break;
-    }
-  }
-
-  boolean hasRemaining() {
-    for (ByteBuffer b : buffers) {
-      if (b.hasRemaining()) return true;
-    }
-    return false;
-  }
-
-  @Override
-  public void close() throws IOException {
-    if (dispose != null) {
-      dispose.close();
-    }
-  }
-}
