+++ google-cloud-storage/src/main/java/com/google/cloud/storage/ResponseContentLifecycleManager.java
@@ -1,0 +1,16 @@
+package com.google.cloud.storage;
+
+import com.google.storage.v2.ReadObjectResponse;
+
+interface ResponseContentLifecycleManager {
+    ResponseContentLifecycleHandle get(ReadObjectResponse response);
+
+    static ResponseContentLifecycleManager noop() {
+        return response ->
+                new ResponseContentLifecycleHandle(
+                        response,
+                        () -> {
+                            // no-op
+                        });
+    }
+}
